move populate to client tests
